separate rule service
