+++ lib/extensions/src/lib/services/extension.service.ts
@@ -16,13 +16,14 @@
  */
 
 import { Injectable, Type } from '@angular/core';
-import { RuleEvaluator, RuleRef, RuleContext, RuleParameter } from '../config/rule.extensions';
+import { RuleEvaluator, RuleRef, RuleContext } from '../config/rule.extensions';
 import { ExtensionConfig } from '../config/extension.config';
 import { ExtensionLoaderService } from './extension-loader.service';
 import { RouteRef } from '../config/routing.extensions';
 import { ActionRef } from '../config/action.extensions';
 import * as core from '../evaluators/core.evaluators';
 import { ComponentRegisterService } from './component-register.service';
+import { RuleService } from './rule.service';
 
 @Injectable({
     providedIn: 'root'
@@ -31,17 +32,15 @@ export class ExtensionService {
     configPath = 'assets/app.extensions.json';
     pluginsPath = 'assets/plugins';
 
-    rules: Array<RuleRef> = [];
     routes: Array<RouteRef> = [];
     actions: Array<ActionRef> = [];
     features: Array<any> = [];
-
     authGuards: { [key: string]: Type<{}> } = {};
-    evaluators: { [key: string]: RuleEvaluator } = {};
 
     constructor(
-        private loader: ExtensionLoaderService,
-        private componentRegister: ComponentRegisterService
+        protected loader: ExtensionLoaderService,
+        protected componentRegister: ComponentRegisterService,
+        protected ruleService: RuleService
     ) {
     }
 
@@ -74,10 +73,11 @@ export class ExtensionService {
             'core.not': core.not
         });
 
-        this.rules = this.loader.getRules(config);
         this.actions = this.loader.getActions(config);
         this.routes = this.loader.getRoutes(config);
         this.features = this.loader.getFeatures(config);
+
+        this.ruleService.setup(config);
     }
 
     /**
@@ -95,9 +95,7 @@ export class ExtensionService {
      * @param values The new evaluators to add
      */
     setEvaluators(values: { [key: string]: RuleEvaluator }) {
-        if (values) {
-            this.evaluators = Object.assign({}, this.evaluators, values);
-        }
+        this.ruleService.setEvaluators(values);
     }
 
     /**
@@ -153,36 +151,17 @@ export class ExtensionService {
      * @returns RuleEvaluator or null if not found
      */
     getEvaluator(key: string): RuleEvaluator {
-        if (key && key.startsWith('!')) {
-            const fn = this.evaluators[key.substring(1)];
-            return (context: RuleContext, ...args: RuleParameter[]): boolean => {
-                return !fn(context, ...args);
-            };
-        }
-        return this.evaluators[key];
+        return this.ruleService.getEvaluator(key);
     }
 
     /**
      * Evaluates a rule.
      * @param ruleId ID of the rule to evaluate
-     * @param context Parameter object for the evaluator with details of app state
+     * @param context (optional) Custom rule execution context.
      * @returns True if the rule passed, false otherwise
      */
-    evaluateRule(ruleId: string, context: RuleContext): boolean {
-        const ruleRef = this.getRuleById(ruleId);
-
-        if (ruleRef) {
-            const evaluator = this.getEvaluator(ruleRef.type);
-            if (evaluator) {
-                return evaluator(context, ...ruleRef.parameters);
-            }
-        } else {
-            const evaluator = this.getEvaluator(ruleId);
-            if (evaluator) {
-                return evaluator(context);
-            }
-        }
-        return false;
+    evaluateRule(ruleId: string, context?: RuleContext): boolean {
+        return this.ruleService.evaluateRule(ruleId, context);
     }
 
     /**
@@ -200,7 +179,7 @@ export class ExtensionService {
      * @returns The rule or null if not found
      */
     getRuleById(id: string): RuleRef {
-        return this.rules.find((ref) => ref.id === id);
+        return this.ruleService.getRuleById(id);
     }
 
     /**
